sink(ticdc): add tests for enable-tidb-extension and extract the config out (#3693)
